Major changes in the system architecture...
So, this was the situation.... My first version did all the necessary tasks in the rubric. Especially, authentication. Authentication was always a mystery  to me and  Google tech enchants me immensely. So, I used Firebase(Google) for authentication. Firebase is used mainly for mobile apps and mobile is another mystery that lured me into using Firebase. On the top of it, Azure is my home, so, I can use AAD anytime. So, I decided to use Firebase. And I did it. My first version was successful for what we targeted for... We just wanted put check marks against the Rubric. And I did checked most of the marks.  Well, my frantic pace yielded all the points on the rubric, however, I found out one huge security flaw in my authentication mechanism! My authentication implementation had many limitations but the biggest was - the first security hole was - The authentication was implemented on the client side.  That meant, I had to keep the authentication secretes on the client side.... Wow! that was a huge security hole! and of course, you can imagine, this is such a debacle, and it indeed opened a really huge can of worms  .... Client side authentication, means,  every time I wanted to make a backend call, I had pass the user credentials...  That would not be acceptable by any security review board. Along with those security holes, that kind-of implementation also creates substandard user experience(session and cookie issues, etc).  Looking at the issues, I decided that I had to move the authentication implementation on the back end. This was a monumental effort.